Moving purely math functions to mathematics.py
