+++ armi/utils/__init__.py
@@ -39,6 +39,7 @@ import armi
 from armi import runLog
 from armi.utils import iterables
 from armi.utils.flags import Flag
+from armi.utils.mathematics import *
 
 # Read in file 1 MB at a time to reduce memory burden of reading entire file at once
 _HASH_BUFFER_SIZE = 1024 * 1024
@@ -239,48 +240,6 @@ def linearInterpolation(x0, y0, x1, y1, targetX=None, targetY=None):
         return m * targetX + b
     else:
         return (targetY - b) / m
-
-
-def parabolaFromPoints(p1, p2, p3):
-    r"""
-    find the parabola that passes through three points
-
-    We solve a simultaneous equation with three points.
-
-    A = x1**2 x1 1
-        x2**2 x2 1
-        x3**2 x3 1
-
-    b = y1
-        y2
-        y3
-
-    find coefficients Ax=b
-
-    Parameters
-    ----------
-    p1 : tuple
-        first point (x,y) coordinates
-    p2,p3: tuple, second and third points.
-
-    Returns
-    -------
-    a,b,c coefficients of y=ax^2+bx+c
-
-    """
-
-    A = numpy.array(
-        [[p1[0] ** 2, p1[0], 1], [p2[0] ** 2, p2[0], 1], [p3[0] ** 2, p3[0], 1]]
-    )
-
-    b = numpy.array([[p1[1]], [p2[1]], [p3[1]]])
-    try:
-        x = numpy.linalg.solve(A, b)
-    except:
-        print("Error in parabola {} {}".format(A, b))
-        raise
-
-    return float(x[0]), float(x[1]), float(x[2])
 
 
 def parabolicInterpolation(ap, bp, cp, targetY):
