+++ doc/conf.py
@@ -61,7 +61,6 @@ armi._ignoreConfigures = True
 
 APIDOC_REL = ".apidocs"
 SOURCE_DIR = os.path.join("..", "armi")
-APIDOC_DIR = APIDOC_REL
 _TUTORIAL_FILES = [
     pathlib.Path(SOURCE_DIR) / "tests" / "tutorials" / fName
     for fName in bookkeepingTests.TUTORIAL_FILES
